backend/requirements.txt
# =============================================================================
# Dépendances Backend — Système de Détection d'Incendies
# Installation: pip install -r requirements.txt
# =============================================================================

# --- API ---
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.1.0

# --- Base de données ---
psycopg2-binary==2.9.11
influxdb-client==1.50.0

# --- Deep Learning ---
torch==2.10.0
numpy==2.4.2

# --- Configuration ---
PyYAML==6.0.3

# --- Data Science (scripts d'entraînement) ---
scikit-learn==1.8.0
matplotlib==3.10.8
seaborn==0.13.2
pandas==3.0.1

# --- Communication série (LoRa) ---
pyserial==3.5

# --- Tests ---
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
